+++ Trivial.Domain/Events/TouchEventArgs.cs
@@ -1,4 +1,4 @@
 ﻿namespace Trivial.Domain.Events;
 
 public record TouchEventArgs(TouchPoint[] ChangedTouches, bool CtrlKey, bool ShiftKey, bool AltKey);
-public record TouchPoint(long Identifier, double ClientX, double ClientY);
+public record TouchPoint(long Identifier, float ClientX, float ClientY);
