+++ Trivial.Domain/Extensions/NumberExtensions.cs
@@ -4,5 +4,5 @@ namespace Trivial.Domain.Extensions;
 
 public static class NumberExtensions
 {
-    public static string ToInvariantString(this double n) => n.ToString(CultureInfo.InvariantCulture);
+    public static string ToInvariantString(this float n) => n.ToString(CultureInfo.InvariantCulture);
 }
