+++ Trivial.Domain/Geometry/Size.cs
@@ -4,12 +4,12 @@ public record Size
 {
     public static Size Zero { get; } = new(0, 0);
 
-    public Size(double width, double height)
+    public Size(float width, float height)
     {
         Width = width;
         Height = height;
     }
 
-    public double Width { get; init; }
-    public double Height { get; init; }
+    public float Width { get; init; }
+    public float Height { get; init; }
 }
