+++ Trivial.Graph/Extensions/EventsExtensions.cs
@@ -7,13 +7,13 @@ public static class EventsExtensions
 {
     public static PointerEventArgs ToCore(this Microsoft.AspNetCore.Components.Web.PointerEventArgs e)
     {
-        return new PointerEventArgs(e.ClientX, e.ClientY, e.Button, e.Buttons, e.CtrlKey, e.ShiftKey, e.AltKey,
+        return new PointerEventArgs((float)e.ClientX, (float)e.ClientY, e.Button, e.Buttons, e.CtrlKey, e.ShiftKey, e.AltKey,
             e.PointerId, e.Width, e.Height, e.Pressure, e.TiltX, e.TiltY, e.PointerType, e.IsPrimary);
     }
 
     public static PointerEventArgs ToCore(this MouseEventArgs e)
     {
-        return new PointerEventArgs(e.ClientX, e.ClientY, e.Button, e.Buttons, e.CtrlKey, e.ShiftKey, e.AltKey,
+        return new PointerEventArgs((float)e.ClientX, (float)e.ClientY, e.Button, e.Buttons, e.CtrlKey, e.ShiftKey, e.AltKey,
             0, 0, 0, 0, 0, 0, string.Empty, false);
     }
 
@@ -24,7 +24,7 @@ public static class EventsExtensions
 
     public static WheelEventArgs ToCore(this Microsoft.AspNetCore.Components.Web.WheelEventArgs e)
     {
-        return new WheelEventArgs(e.ClientX, e.ClientY, e.Button, e.Buttons, e.CtrlKey, e.ShiftKey, e.AltKey, e.DeltaX,
-            e.DeltaY, e.DeltaZ, e.DeltaMode);
+        return new WheelEventArgs((float)e.ClientX, (float)e.ClientY, e.Button, e.Buttons, e.CtrlKey, e.ShiftKey, e.AltKey, (float)e.DeltaX,
+            (float)e.DeltaY, (float)e.DeltaZ, e.DeltaMode);
     }
 }
